+++ src/Build5Nines.SharpVector/VectorTextResult.cs
@@ -3,10 +3,11 @@ namespace Build5Nines.SharpVector;
 
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.InteropServices;
 
-public interface IVectorTextResult<TDocument, TMetadata>
+public interface IVectorTextResult<TId, TDocument, TMetadata>
 {
-    IEnumerable<IVectorTextResultItem<TDocument, TMetadata>> Texts { get; }
+    IEnumerable<IVectorTextResultItem<TId, TDocument, TMetadata>> Texts { get; }
 
     /// <summary>
     /// Returns true if the search returned no results.
@@ -30,12 +31,13 @@ public interface IVectorTextResult<TDocument, TMetadata>
 }
 
 public interface IVectorTextResult<TMetadata>
- : IVectorTextResult<string, TMetadata>
+ : IVectorTextResult<int, string, TMetadata>
  { }
 
-public class VectorTextResult<TDocument, TMetadata> : IVectorTextResult<TDocument, TMetadata>
+public class VectorTextResult<TId, TDocument, TMetadata>
+    : IVectorTextResult<TId, TDocument, TMetadata>
 {
-    public VectorTextResult(int totalCount, int pageIndex, int totalPages, IEnumerable<IVectorTextResultItem<TDocument, TMetadata>> texts)
+    public VectorTextResult(int totalCount, int pageIndex, int totalPages, IEnumerable<IVectorTextResultItem<TId, TDocument, TMetadata>> texts)
     {
         Texts = texts;
         TotalCount = totalCount;
@@ -46,7 +48,7 @@ public class VectorTextResult<TDocument, TMetadata> : IVectorTextResult<TDocumen
     /// <summary>
     /// Returns true if the search returned no results.
     /// </summary>
-    public IEnumerable<IVectorTextResultItem<TDocument, TMetadata>> Texts { get; private set; }
+    public IEnumerable<IVectorTextResultItem<TId, TDocument, TMetadata>> Texts { get; private set; }
 
     public bool IsEmpty { get => Texts == null || !Texts.Any(); }
 
@@ -66,9 +68,10 @@ public class VectorTextResult<TDocument, TMetadata> : IVectorTextResult<TDocumen
     public int TotalPages { get; private set; }
 }
 
-public class VectorTextResult<TMetadata> : VectorTextResult<string, TMetadata>, IVectorTextResult<TMetadata>
+public class VectorTextResult<TMetadata>
+    : VectorTextResult<int, string, TMetadata>, IVectorTextResult<TMetadata>
 {
-    public VectorTextResult(int totalCount, int pageIndex, int totalPages, IEnumerable<IVectorTextResultItem<string, TMetadata>> texts)
+    public VectorTextResult(int totalCount, int pageIndex, int totalPages, IEnumerable<IVectorTextResultItem<int, string, TMetadata>> texts)
         : base(totalCount, pageIndex, totalPages, texts)
     { }
 }